+++ lib/integration/marketpress.inc.php
@@ -214,7 +214,6 @@ class MP_Gateway_Paymill_for_WordPress extends MP_Gateway_API {
 			paymill_form_checkout_submit_id = "'.$submit_id.'";
 			paymill_shop_name = "marketpress";
 			paymill_pcidss3 = '.((empty($GLOBALS['paymill_settings']->paymill_general_settings['pci_dss_3']) || $GLOBALS['paymill_settings']->paymill_general_settings['pci_dss_3'] != '1') ? 1 : 0).';
-			paymill_pcidss3_lang = "'.substr(apply_filters('plugin_locale', get_locale(), $domain),0,2).'";
 			</script>';
 			
 			echo do_shortcode($mp->get_setting('gateways->paymill-for-wordpress->instructions'));