+++ lib/integration/pay_button.inc.php
@@ -318,7 +318,6 @@
 					paymill_form_checkout_submit_id = "'.$submit_id.'";
 					paymill_shop_name = "paybutton";
 					paymill_pcidss3 = '.((empty($GLOBALS['paymill_settings']->paymill_general_settings['pci_dss_3']) || $GLOBALS['paymill_settings']->paymill_general_settings['pci_dss_3'] != '1') ? 1 : 0).';
-					paymill_pcidss3_lang = "'.substr(apply_filters('plugin_locale', get_locale(), $domain),0,2).'";
 					</script>';
 					
 					if($this->subscriptions === false){
@@ -365,12 +364,12 @@
 			return $instance;
 		}
 		function form($instance) {
-			$products_whitelist = unserialize($instance['products']);
+			$products_whitelist = (isset($instance['products']) ? unserialize($instance['products']) : false);
 			echo'
 			<fieldset>
 				<legend><h4>'.__('Title:', 'paymill').'</h4></legend>
 				<label for="'.$this->get_field_id('title').'">
-					<input class="widefat" id="'.$this->get_field_id('title').'" name="'.$this->get_field_name('title').'" type="text" value="'.esc_attr($instance['title']).'" />
+					<input class="widefat" id="'.$this->get_field_id('title').'" name="'.$this->get_field_name('title').'" type="text" value="'.esc_attr((isset($instance['title']) ? $instance['title'] : '')).'" />
 				</label>
 			</fieldset>
 			<br />